requirements.txt
PyYAML~=6.0
biliup ~= 0.4.17